fix(db): re-type mbbo/mbbi raw/mask/state fields to DBF_ULONG/USHORT
mbboRecord.dbd.pod / mbbiRecord.dbd.pod declare the raw, mask and
state-table fields as unsigned, but the Rust records modeled them as
signed Long/Short/Enum, so a CA read advertised the wrong DBR type
(DBR_LONG instead of DBR_DOUBLE for DBF_ULONG, DBR_LONG vs DBR_LONG-from
-USHORT) and a raw/mask value with the high bit set (>= 2^31) was stored
in i32 and lost its sign on round-trip.

Re-type descriptors AND storage:
- mbbo: RVAL/ORAW/RBV/ORBV/MASK and the 16 state values ZRVL..FFVL ->
  DBF_ULONG (u32); NOBT/SHFT/MLST/LALM/IVOV -> DBF_USHORT (u16).
- mbbi: RVAL/ORAW/MASK and ZRVL..FFVL -> DBF_ULONG (u32);
  NOBT/SHFT/MLST/LALM -> DBF_USHORT (u16).

get_field now serves the native unsigned variant (correct wire type).
The mbb_put_field! macro gains a per-variant tolerant coercion encoded
once: a DBF_ULONG field accepts EpicsValue::ULong natively and tolerates
the legacy signed Long carrier; a DBF_USHORT field accepts UShort and
tolerates Enum/Short. The signed->unsigned reinterpret preserves the bit
pattern, so device support / autosave callers (e.g. the asynUInt32Digital
MASK/SHFT propagation in asyn-rs) keep working without change.

mbbi.SVAL (mbbiRecord.dbd.pod:663, DBF_ULONG) is not modeled in the Rust
mbbi (no field, no descriptor), so there is no site to re-type.

Tests asserting the old signed wire variant for a genuinely-unsigned
field are updated to the correct unsigned variant, each citing the C
.dbd line; a high-bit (0x8000_0000) round-trip test is added for
mbbo.ZRVL through convert().

Closes R0604-BASEDB-UNSIGNED-FIELD-TYPES-1 (part b) for the mbbo/mbbi
record family. Ref: epics-base modules/database/src/std/rec/
mbboRecord.dbd.pod, mbbiRecord.dbd.pod.

+++ crates/epics-base-rs/tests/c_epics_parity.rs
@@ -1176,13 +1176,15 @@ fn mbbo_state_strings_and_values() {
     rec.put_field("TWST", EpicsValue::String("High".into()))
         .unwrap();
 
-    // C defaults all *VL to 0. Set them explicitly.
+    // C defaults all *VL to 0. Set them explicitly. ZRVL..FFVL are
+    // DBF_ULONG (mbboRecord.dbd.pod:222-342); the legacy signed Long put is
+    // still tolerated, but the read-back is the native ULong.
     rec.put_field("ZRVL", EpicsValue::Long(0)).unwrap();
     rec.put_field("ONVL", EpicsValue::Long(1)).unwrap();
     rec.put_field("TWVL", EpicsValue::Long(2)).unwrap();
-    assert_eq!(rec.get_field("ZRVL"), Some(EpicsValue::Long(0)));
-    assert_eq!(rec.get_field("ONVL"), Some(EpicsValue::Long(1)));
-    assert_eq!(rec.get_field("TWVL"), Some(EpicsValue::Long(2)));
+    assert_eq!(rec.get_field("ZRVL"), Some(EpicsValue::ULong(0)));
+    assert_eq!(rec.get_field("ONVL"), Some(EpicsValue::ULong(1)));
+    assert_eq!(rec.get_field("TWVL"), Some(EpicsValue::ULong(2)));
 
     rec.put_field("VAL", EpicsValue::Enum(1)).unwrap();
     assert_eq!(rec.get_field("VAL"), Some(EpicsValue::Enum(1)));
@@ -1220,9 +1222,11 @@ fn mbbi_all_16_states() {
             15 => "FFVL",
             _ => unreachable!(),
         };
+        // ZRVL..FFVL are DBF_ULONG (mbbiRecord.dbd.pod:144-264); served as
+        // the native ULong even though the put used the tolerated Long.
         assert_eq!(
             rec.get_field(field),
-            Some(EpicsValue::Long(i as i32)),
+            Some(EpicsValue::ULong(u32::from(i))),
             "State {i} value mismatch"
         );
     }
